--- lib/Doctrine/ORM/Mapping/AssociationOwningSideMapping.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Doctrine\ORM\Mapping;
-
-interface AssociationOwningSideMapping
-{
-}
