+++ lib/Doctrine/ORM/Mapping/ManyToOneAssociationMapping.php
@@ -7,26 +7,6 @@ namespace Doctrine\ORM\Mapping;
 /**
  * The "many" side of a many-to-one association mapping is always the owning side.
  */
-final class ManyToOneAssociationMapping extends ToOneAssociationMapping implements AssociationOwningSideMapping
+final class ManyToOneAssociationMapping extends ToOneOwningSideMapping
 {
-    /** @var array<string, string> */
-    public array $sourceToTargetKeyColumns = [];
-
-    /** @var array<string, string> */
-    public array $targetToSourceKeyColumns = [];
-
-    /** @var list<JoinColumnMapping> */
-    public array $joinColumns = [];
-
-    /** @return list<string> */
-    public function __sleep(): array
-    {
-        $serialized = parent::__sleep();
-
-        $serialized[] = 'joinColumns';
-        $serialized[] = 'sourceToTargetKeyColumns';
-        $serialized[] = 'targetToSourceKeyColumns';
-
-        return $serialized;
-    }
 }
