+++ lib/Doctrine/ORM/Mapping/OneToOneAssociationMapping.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Doctrine\ORM\Mapping;
 
-class OneToOneAssociationMapping extends ToOneAssociationMapping
+interface OneToOneAssociationMapping extends ToOneAssociationMapping
 {
 }
