+++ lib/Doctrine/ORM/Mapping/OneToOneInverseSideMapping.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\ORM\Mapping;
+
+final class OneToOneInverseSideMapping extends ToOneInverseSideMapping implements OneToOneAssociationMapping
+{
+}
