+++ lib/Doctrine/ORM/Mapping/ToManyInverseSideMapping.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\ORM\Mapping;
+
+abstract class ToManyInverseSideMapping extends InverseSideMapping implements ToManyAssociationMapping
+{
+    use ToManyAssociationMappingImplementation;
+}
