+++ lib/Doctrine/ORM/Mapping/ToManyOwningSideMapping.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Doctrine\ORM\Mapping;
+
+abstract class ToManyOwningSideMapping extends OwningSideMapping
+{
+    use ToManyAssociationMappingImplementation;
+}
